Move main file into lib directory.
