+++ lib/index.js
@@ -2,8 +2,7 @@ const through = require('through2');
 const embedJson = require('embed-json');
 const PluginError = require('plugin-error');
 const fs = require('fs');
-const path = require('path');
-const PLUGIN_NAME = require('./package').name;
+const PLUGIN_NAME = require('../package').name;
 const defaults = {
   mimeTypes: [
     'application/json',
